perf: switch to bf16 training precision — 4x decoder speedup on tensor cores
- precision: float32 -> bfloat16 in run_monitored_training.py
  (entire model stack now uses bf16 tensor cores for GEMMs)
- _ieee_fp32_math decorator now also disables autocast, protecting
  the LBM solver from ambient bf16 casting (solver stays IEEE fp32)

Impact at 128^3 w=1600 d=12 on L4:
- Decoder FLOPs unchanged (~960 TFLOP/update from full-lattice loss)
- But throughput: ~30 TFLOPS fp32 -> ~92 TFLOPS bf16 = 3x faster decoder
- Combined with solver interval=4 and torch.compile:
  per-update drops from ~92s to potentially ~15-20s

+++ CLI/run_monitored_training.py
@@ -1188,7 +1188,7 @@ def main() -> int:
         batch_size=args.batch_size,
         learning_rate=args.learning_rate,
         disconnection_penalty=30.0,
-        precision="float32",
+        precision="bfloat16",
         enable_pipeline_parallelism=False,
         direct_solver_loss_weight=args.direct_solver_loss_weight,
         direct_solver_interval=int(config_value("training", "direct_solver_interval", 4)),
@@ -1693,4 +1693,5 @@ def main() -> int:
 
 if __name__ == "__main__":
     raise SystemExit(main())
+
